requirements.txt
selenium>=4.0.0
Pillow>=10.0.0
webdriver-manager>=4.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
matplotlib>=3.5.0
reportlab>=3.6.0